+++ test/web-api.test.js
@@ -3,85 +3,9 @@ const crypto = require('crypto');
 const keypair = require('keypair');
 const nock = require('nock');
 
-const SmileIdentity = require('..');
-
-const WebApi = SmileIdentity.WebApi;
-const Signature = SmileIdentity.Signature;
-const Utilities = SmileIdentity.Utilities;
-const IDApi = SmileIdentity.IDApi;
+const { WebApi, Signature } = require('..');
 
 const pair = keypair();
-
-// test that the sec key is generated correctly
-describe('Signature', function () {
-  describe('#new', function () {
-    it('should set the partner_id and api_key values', function (done) {
-      const instance = new Signature('001', Buffer.from(pair.public).toString('base64'));
-      assert.equal(instance.partnerID, '001');
-      assert.equal(instance.apiKey, Buffer.from(pair.public).toString('base64'));
-      done();
-    });
-  });
-
-  describe('#generate_sec_key', function () {
-    it('should create a sec_key', function (done) {
-      const timestamp = Date.now();
-      const signature = new Signature('001', Buffer.from(pair.public).toString('base64')).generate_sec_key(timestamp);
-      assert.equal(typeof (signature), 'object');
-      assert.equal(timestamp, signature.timestamp);
-      const hash = crypto.createHash('sha256').update(`${1}:${timestamp}`).digest('hex');
-      assert.equal(hash, signature.sec_key.split('|')[1]);
-      const decrypted = crypto.privateDecrypt({
-        key: Buffer.from(pair.private),
-        padding: crypto.constants.RSA_PKCS1_PADDING,
-      }, Buffer.from(signature.sec_key.split('|')[0], 'base64')).toString();
-      assert.equal(decrypted, hash);
-      done();
-    });
-  });
-
-  describe('#confirm_sec_key', function () {
-    it('should be able to decode a valid sec_key', function (done) {
-      const timestamp = Date.now();
-      const hash = crypto.createHash('sha256').update(`${1}:${timestamp}`).digest('hex');
-      const encrypted = crypto.privateEncrypt({
-        key: Buffer.from(pair.private),
-        padding: crypto.constants.RSA_PKCS1_PADDING,
-      }, Buffer.from(hash)).toString('base64');
-      const sec_key = [encrypted, hash].join('|');
-      assert.equal(true, new Signature('001', Buffer.from(pair.public).toString('base64')).confirm_sec_key(timestamp, sec_key));
-      done();
-    });
-  });
-
-  describe('#generate_signature', function () {
-    it('should calculate a signature and use a timestamp if provided one', function (done) {
-      const timestamp = new Date().toISOString();
-      const hmac = crypto.createHmac('sha256', '1234');
-      hmac.update(timestamp, 'utf8');
-      hmac.update('002', 'utf8');
-      hmac.update('sid_request', 'utf8');
-      const output = hmac.digest().toString('base64');
-      const result = new Signature('002', '1234').generate_signature(timestamp);
-      assert.equal(output, result.signature);
-      assert.equal(timestamp, result.timestamp);
-      done();
-    });
-  });
-
-  describe('#confirm_signature', function () {
-    it('should confirm an incoming signaute', function (done) {
-      const timestamp = new Date().toISOString();
-      const hmac = crypto.createHmac('sha256', '1234');
-      hmac.update(timestamp, 'utf8');
-      hmac.update('002', 'utf8');
-      hmac.update('sid_request', 'utf8');
-      const output = hmac.digest().toString('base64');
-      assert.equal(true, new Signature('002', '1234').confirm_signature(timestamp, output));
-      done();
-    });
-  });
-});
 
 describe('WebApi', function () {
   describe('#new', function () {
@@ -1008,432 +932,6 @@ describe('WebApi', function () {
           assert.equal(resp.token, '42');
           done();
         });
-      });
-    });
-  });
-});
-
-describe('Utilities', function () {
-  describe('#get_job_status', function () {
-    it('should be able to check job_status successfully', function (done) {
-      const partner_params = {
-        user_id: '1',
-        job_id: '1',
-        job_type: 4,
-      };
-      const options = {
-        return_images: true,
-        return_history: true,
-      };
-
-      const timestamp = Date.now();
-      const hash = crypto.createHash('sha256').update(`${1}:${timestamp}`).digest('hex');
-      const encrypted = crypto.privateEncrypt({
-        key: Buffer.from(pair.private),
-        padding: crypto.constants.RSA_PKCS1_PADDING,
-      }, Buffer.from(hash)).toString('base64');
-      const sec_key = [encrypted, hash].join('|');
-      const jobStatusResponse = {
-        job_success: true,
-        job_complete: true,
-        result: {
-          ResultCode: '0810',
-          ResultText: 'Awesome!',
-        },
-        timestamp: timestamp,
-        signature: sec_key,
-      };
-      nock('https://testapi.smileidentity.com')
-        .post('/v1/job_status', (body) => {
-          assert.equal(body.job_id, partner_params.job_id);
-          assert.equal(body.user_id, partner_params.user_id);
-          assert.notEqual(body.timestamp, undefined);
-          assert.notEqual(body.sec_key, undefined);
-          assert.equal(body.image_links, true);
-          assert.equal(body.history, true);
-          return true;
-        })
-        .reply(200, jobStatusResponse)
-        .isDone();
-      new Utilities('001', Buffer.from(pair.public).toString('base64'), 0)
-        .get_job_status(partner_params.user_id, partner_params.job_id, options)
-        .then((job_status, err) => {
-          assert.equal(job_status.sec_key, jobStatusResponse.sec_key);
-          assert.equal(job_status.job_complete, true);
-          done();
-        }).catch((err) => {
-          assert.equal(null, err);
-          console.log(err);
-        });
-    });
-
-    it('should be able to use the signature instead of the sec_key when provided an option flag', function (done) {
-      const partner_params = {
-        user_id: '1',
-        job_id: '1',
-        job_type: 4,
-      };
-      const options = {
-        return_images: true,
-        return_history: true,
-        signature: true,
-      };
-
-      const timestamp = new Date().toISOString();
-      const signature = new Signature('001', '1234').generate_signature(timestamp).signature;
-
-      const jobStatusResponse = {
-        job_success: true,
-        job_complete: true,
-        result: {
-          ResultCode: '0810',
-          ResultText: 'Awesome!',
-        },
-        timestamp: timestamp,
-        signature: signature,
-      };
-      nock('https://testapi.smileidentity.com')
-        .post('/v1/job_status', (body) => {
-          assert.equal(body.job_id, partner_params.job_id);
-          assert.equal(body.user_id, partner_params.user_id);
-          assert.notEqual(body.timestamp, undefined);
-          assert.notEqual(body.signature, undefined);
-          assert.equal(body.image_links, true);
-          assert.equal(body.history, true);
-          return true;
-        })
-        .reply(200, jobStatusResponse)
-        .isDone();
-      new Utilities('001', '1234', 0)
-        .get_job_status(partner_params.user_id, partner_params.job_id, options)
-        .then((job_status, err) => {
-          assert.equal(job_status.signature, jobStatusResponse.signature);
-          assert.equal(job_status.job_complete, true);
-          done();
-        }).catch((err) => {
-          assert.equal(null, err);
-          console.log(err);
-        });
-    });
-
-    it('should raise an error if one occurs', function (done) {
-      const partner_params = {
-        user_id: '1',
-        job_id: '1',
-        job_type: 4,
-      };
-      const options = {
-        return_images: true,
-        return_history: true,
-      };
-
-      const timestamp = Date.now();
-      const hash = crypto.createHash('sha256').update(`${1}:${timestamp}`).digest('hex');
-      const encrypted = crypto.privateEncrypt({
-        key: Buffer.from(pair.private),
-        padding: crypto.constants.RSA_PKCS1_PADDING,
-      }, Buffer.from(hash)).toString('base64');
-      const sec_key = [encrypted, hash].join('|');
-      const jobStatusResponse = {
-        error: 'oops',
-      };
-      nock('https://testapi.smileidentity.com')
-        .post('/v1/job_status', (body) => {
-          assert.equal(body.job_id, partner_params.job_id);
-          assert.equal(body.user_id, partner_params.user_id);
-          assert.notEqual(body.timestamp, undefined);
-          assert.notEqual(body.sec_key, undefined);
-          assert.equal(body.image_links, true);
-          assert.equal(body.history, true);
-          return true;
-        })
-        .replyWithError(400, {
-          code: '2204',
-          error: 'unauthorized',
-        })
-        .isDone();
-      new Utilities('001', Buffer.from(pair.public).toString('base64'), 0)
-        .get_job_status(partner_params.user_id, partner_params.job_id, options)
-        .then((job_status) => {
-          assert.equal(null, job_status);
-        }).catch((err) => {
-          assert.equal(err.message, 'Error: 400');
-          done();
-        });
-    });
-  });
-});
-
-describe('IDapi', function () {
-  describe('#new', function () {
-    it('should instantiate and set the global variables', function (done) {
-      const instance = new IDApi('001', Buffer.from(pair.public).toString('base64'), 0);
-      assert.equal(instance.partner_id, '001');
-      assert.equal(instance.api_key, Buffer.from(pair.public).toString('base64'));
-      assert.equal(instance.url, 'testapi.smileidentity.com/v1');
-      done();
-    });
-  });
-
-  describe('#submit_job', function () {
-    it('should ensure that the partner_params are present', function (done) {
-      const instance = new IDApi('001', Buffer.from(pair.public).toString('base64'), 0);
-      instance.submit_job(null, {}).catch((err) => {
-        assert.equal(err.message, 'Please ensure that you send through partner params');
-        done();
-      });
-    });
-
-    it('should ensure that the partner_params are an object', function (done) {
-      const instance = new IDApi('001', Buffer.from(pair.public).toString('base64'), 0);
-      instance.submit_job('not partner params', {}).catch((err) => {
-        assert.equal(err.message, 'Partner params needs to be an object');
-        done();
-      });
-    });
-
-    it('should ensure that the partner_params contain user_id, job_id and job_type', function (done) {
-      const instance = new IDApi('001', Buffer.from(pair.public).toString('base64'), 0);
-      ['user_id', 'job_id', 'job_type'].forEach((key) => {
-        const partner_params = {
-          user_id: '1',
-          job_id: '1',
-          job_type: 5,
-        };
-        delete partner_params[key];
-        instance.submit_job(partner_params, {}, {}, { return_job_status: true }).catch((err) => {
-          assert.equal(err.message, `Please make sure that ${key} is included in the partner params`);
-        });
-      });
-      done();
-    });
-
-    it('should ensure that in partner_params, user_id, job_id, and job_type are not emptystrings', function (done) {
-      const instance = new IDApi('001', Buffer.from(pair.public).toString('base64'), 0);
-      ['user_id', 'job_id', 'job_type'].forEach((key) => {
-        const partner_params = {
-          user_id: '1',
-          job_id: '1',
-          job_type: 5,
-        };
-        partner_params[key] = '';
-        instance.submit_job(partner_params, {}, {}, { return_job_status: true }).catch((err) => {
-          assert.equal(err.message, `Please make sure that ${key} is included in the partner params`);
-        });
-      });
-      done();
-    });
-
-    it('should ensure that the id_info is an object', function (done) {
-      const instance = new IDApi('001', Buffer.from(pair.public).toString('base64'), 0);
-      instance.submit_job({ user_id: '1', job_id: '1', job_type: 5 }, '').catch((err) => {
-        assert.equal(err.message, 'ID Info needs to be an object');
-        done();
-      });
-    });
-
-    it('should ensure that the id_info object is not empty or nil', function (done) {
-      const instance = new IDApi('001', Buffer.from(pair.public).toString('base64'), 0);
-      instance.submit_job({ user_id: '1', job_id: '1', job_type: 5 }, {}).catch((err) => {
-        assert.equal(err.message, 'Please make sure that id_info not empty or nil');
-        done();
-      });
-    });
-
-    it('should ensure that the id_info object is not empty or nil', function (done) {
-      const instance = new IDApi('001', Buffer.from(pair.public).toString('base64'), 0);
-      instance.submit_job({ user_id: '1', job_id: '1', job_type: 5 }, { id_number: '' }).catch((err) => {
-        assert.equal(err.message, 'Please provide an id_number in the id_info payload');
-        done();
-      });
-    });
-
-    it('should ensure that the the job id is set to 5', function (done) {
-      const instance = new IDApi('001', Buffer.from(pair.public).toString('base64'), 0);
-      const partner_params = {
-        user_id: '1',
-        job_id: '1',
-        job_type: 4,
-      };
-      instance.submit_job(partner_params, null).catch((err) => {
-        assert.equal(err.message, 'Please ensure that you are setting your job_type to 5 to query ID Api');
-        done();
-      });
-    });
-
-    it('should be able to send a job', function (done) {
-      const instance = new IDApi('001', Buffer.from(pair.public).toString('base64'), 0);
-      const partner_params = {
-        user_id: '1',
-        job_id: '1',
-        job_type: 5,
-      };
-      const id_info = {
-        first_name: 'John',
-        last_name: 'Doe',
-        middle_name: '',
-        country: 'NG',
-        id_type: 'BVN',
-        id_number: '00000000000',
-        phone_number: '0726789065',
-      };
-      const IDApiResponse = {
-        JSONVersion: '1.0.0',
-        SmileJobID: '0000001096',
-        PartnerParams: {
-          user_id: 'dmKaJazQCziLc6Tw9lwcgzLo',
-          job_id: 'DeXyJOGtaACFFfbZ2kxjuICE',
-          job_type: 5,
-        },
-        ResultType: 'ID Verification',
-        ResultText: 'ID Number Validated',
-        ResultCode: '1012',
-        IsFinalResult: 'true',
-        Actions: {
-          Verify_ID_Number: 'Verified',
-          Return_Personal_Info: 'Returned',
-        },
-        Country: 'NG',
-        IDType: 'BVN',
-        IDNumber: '00000000000',
-        ExpirationDate: 'NaN-NaN-NaN',
-        FullName: 'some  person',
-        DOB: 'NaN-NaN-NaN',
-        Photo: 'Not Available',
-        sec_key: 'RKYX2ZVpvNTFW8oXdN3iTvQcefV93VMo18LQ/Uco0=|7f0b0d5ebc3e5499c224f2db478e210d1860f01368ebc045c7bbe6969f1c08ba',
-        timestamp: 1570612182124,
-      };
-
-      nock('https://testapi.smileidentity.com')
-        .post('/v1/id_verification', (body) => {
-          assert.equal(body.partner_id, '001');
-          assert.notEqual(body.sec_key, undefined);
-          assert.notEqual(body.timestamp, undefined);
-          assert.equal(body.partner_params.user_id, partner_params.user_id);
-          assert.equal(body.partner_params.job_id, partner_params.job_id);
-          assert.equal(body.partner_params.job_type, partner_params.job_type);
-          assert.equal(body.first_name, id_info.first_name);
-          assert.equal(body.last_name, id_info.last_name);
-          assert.equal(body.middle_name, id_info.middle_name);
-          assert.equal(body.country, id_info.country);
-          assert.equal(body.id_type, id_info.id_type);
-          assert.equal(body.id_number, id_info.id_number);
-          assert.equal(body.phone_number, id_info.phone_number);
-          return true;
-        })
-        .reply(200, IDApiResponse)
-        .isDone();
-
-      const promise = instance.submit_job(partner_params, id_info);
-      promise.then((resp) => {
-        assert.deepEqual(Object.keys(resp).sort(), ['JSONVersion', 'SmileJobID', 'PartnerParams', 'ResultType', 'ResultText', 'ResultCode', 'IsFinalResult', 'Actions', 'Country', 'IDType', 'IDNumber', 'ExpirationDate', 'FullName', 'DOB', 'Photo', 'sec_key', 'timestamp'].sort());
-        done();
-      });
-    });
-
-    it('should raise an error when a network call fails', function (done) {
-      const instance = new IDApi('001', Buffer.from(pair.public).toString('base64'), 0);
-      const partner_params = {
-        user_id: '1',
-        job_id: '1',
-        job_type: 5,
-      };
-      const id_info = {
-        first_name: 'John',
-        last_name: 'Doe',
-        middle_name: '',
-        country: 'NG',
-        id_type: 'BVN',
-        id_number: '00000000000',
-        phone_number: '0726789065',
-      };
-
-      nock('https://testapi.smileidentity.com')
-        .post('/v1/id_verification')
-        .replyWithError(400, {
-          code: '2204',
-          error: 'unauthorized',
-        })
-        .isDone();
-
-      instance.submit_job(partner_params, id_info).then((resp) => {
-        assert.equal(false);
-      }).catch((err) => {
-        // todo: figure out how to get nook to act like an error response would in real life
-        // err.message in this case should be '2204:unauthorized'
-        assert.equal(err.message, undefined);
-      });
-
-      done();
-    });
-
-    it('should use the signature instead of sec_key when provided an optional parameter', function (done) {
-      const instance = new IDApi('001', '1234', 0);
-      const partner_params = {
-        user_id: '1',
-        job_id: '1',
-        job_type: 5,
-      };
-      const id_info = {
-        first_name: 'John',
-        last_name: 'Doe',
-        middle_name: '',
-        country: 'NG',
-        id_type: 'BVN',
-        id_number: '00000000000',
-        phone_number: '0726789065',
-      };
-      const IDApiResponse = {
-        JSONVersion: '1.0.0',
-        SmileJobID: '0000001096',
-        PartnerParams: {
-          user_id: 'dmKaJazQCziLc6Tw9lwcgzLo',
-          job_id: 'DeXyJOGtaACFFfbZ2kxjuICE',
-          job_type: 5,
-        },
-        ResultType: 'ID Verification',
-        ResultText: 'ID Number Validated',
-        ResultCode: '1012',
-        IsFinalResult: 'true',
-        Actions: {
-          Verify_ID_Number: 'Verified',
-          Return_Personal_Info: 'Returned',
-        },
-        Country: 'NG',
-        IDType: 'BVN',
-        IDNumber: '00000000000',
-        ExpirationDate: 'NaN-NaN-NaN',
-        FullName: 'some  person',
-        DOB: 'NaN-NaN-NaN',
-        Photo: 'Not Available',
-        signature: 'RKYX2ZVpvNTFW8oXdN3iTvQcefV93VMo18LQ/Uco0=',
-        timestamp: 1570612182124,
-      };
-
-      nock('https://testapi.smileidentity.com')
-        .post('/v1/id_verification', (body) => {
-          assert.equal(body.partner_id, '001');
-          assert.notEqual(body.signature, undefined);
-          assert.notEqual(body.timestamp, undefined);
-          assert.equal(body.partner_params.user_id, partner_params.user_id);
-          assert.equal(body.partner_params.job_id, partner_params.job_id);
-          assert.equal(body.partner_params.job_type, partner_params.job_type);
-          assert.equal(body.first_name, id_info.first_name);
-          assert.equal(body.last_name, id_info.last_name);
-          assert.equal(body.middle_name, id_info.middle_name);
-          assert.equal(body.country, id_info.country);
-          assert.equal(body.id_type, id_info.id_type);
-          assert.equal(body.id_number, id_info.id_number);
-          assert.equal(body.phone_number, id_info.phone_number);
-          return true;
-        })
-        .reply(200, IDApiResponse)
-        .isDone();
-
-      const promise = instance.submit_job(partner_params, id_info, { signature: true });
-      promise.then((resp) => {
-        assert.deepEqual(Object.keys(resp).sort(), ['JSONVersion', 'SmileJobID', 'PartnerParams', 'ResultType', 'ResultText', 'ResultCode', 'IsFinalResult', 'Actions', 'Country', 'IDType', 'IDNumber', 'ExpirationDate', 'FullName', 'DOB', 'Photo', 'signature', 'timestamp'].sort());
-        done();
       });
     });
   });
